Refactored packet creation and handling into MeshDevice
