+++ src-tauri/src/main.rs
@@ -5,7 +5,6 @@ mod graph;
 mod mesh;
 mod mocks;
 
-use app::protobufs;
 use mesh::serial_connection::{MeshConnection, SerialConnection};
 use std::sync::Arc;
 use tauri::{async_runtime, Manager};
@@ -53,13 +52,15 @@ async fn connect_to_serial_port(
     mesh_device: tauri::State<'_, ActiveMeshDevice>,
     serial_connection: tauri::State<'_, ActiveSerialConnection>,
 ) -> Result<(), String> {
-    let mut connection = SerialConnection::new(SerialConnection::generate_rand_id());
+    let mut connection = SerialConnection::new();
+    let new_device = mesh::device::MeshDevice::new();
+
     connection
         .connect(port_name, 115_200)
         .expect("Could not connect to serial port at 115_200 baud");
 
     connection
-        .configure(connection.config_id)
+        .configure(new_device.config_id)
         .expect("Could not configure serial device");
 
     let mut decoded_listener = connection
@@ -72,8 +73,8 @@ async fn connect_to_serial_port(
     let mesh_device_arc = mesh_device.inner.clone();
 
     {
-        let mut new_device = mesh_device_arc.lock().await;
-        *new_device = Some(mesh::device::MeshDevice::new());
+        let mut new_device_guard = mesh_device_arc.lock().await;
+        *new_device_guard = Some(new_device);
     }
 
     tauri::async_runtime::spawn(async move {
@@ -94,14 +95,13 @@ async fn connect_to_serial_port(
                     }
                 };
 
-                let device_updated =
-                    match SerialConnection::handle_packet_from_radio(variant, device).await {
-                        Ok(d) => d,
-                        Err(e) => {
-                            eprintln!("Error transmitting packet: {}", e.to_string());
-                            continue;
-                        }
-                    };
+                let device_updated = match device.handle_packet_from_radio(variant) {
+                    Ok(d) => d,
+                    Err(e) => {
+                        eprintln!("Error transmitting packet: {}", e.to_string());
+                        continue;
+                    }
+                };
 
                 if device_updated {
                     match dispatch_updated_device(handle.clone(), device.clone()) {
@@ -157,32 +157,28 @@ async fn send_text(
     mesh_device: tauri::State<'_, ActiveMeshDevice>,
     serial_connection: tauri::State<'_, ActiveSerialConnection>,
 ) -> Result<(), String> {
-    let mut guard = serial_connection.inner.lock().await;
-    let connection = guard.as_mut().expect("Connection not initialized");
-    connection
-        .send_text(
-            text.clone(),
-            mesh::serial_connection::PacketDestination::BROADCAST,
-            true,
-            0,
-        )
-        .map_err(|e| e.to_string())?;
-
+    let mut serial_guard = serial_connection.inner.lock().await;
     let mut device_guard = mesh_device.inner.lock().await;
+
+    let connection = serial_guard
+        .as_mut()
+        .ok_or("Connection not initialized")
+        .map_err(|e| e.to_string())?;
 
     let device = device_guard
         .as_mut()
         .ok_or("Device not connected")
         .map_err(|e| e.to_string())?;
 
-    device.add_message(mesh::device::TextPacket {
-        packet: protobufs::MeshPacket {
-            rx_time: mesh::device::helpers::get_current_time_u32(),
-            channel,
-            ..Default::default()
-        },
-        data: text,
-    });
+    device
+        .send_text(
+            connection,
+            text.clone(),
+            mesh::serial_connection::PacketDestination::BROADCAST,
+            true,
+            0,
+        )
+        .map_err(|e| e.to_string())?;
 
     dispatch_updated_device(app_handle, device.clone()).map_err(|e| e.to_string())?;
