+++ src-tauri/src/mesh/device/helpers.rs
@@ -1,3 +1,4 @@
+use rand::{distributions::Standard, prelude::Distribution, Rng};
 use std::time::UNIX_EPOCH;
 
 pub fn get_current_time_u32() -> u32 {
@@ -7,4 +8,12 @@ pub fn get_current_time_u32() -> u32 {
         .as_secs()
         .try_into()
         .expect("Could not convert u128 to u32")
+}
+
+pub fn generate_rand_id<T>() -> T
+where
+    Standard: Distribution<T>,
+{
+    let mut rng = rand::thread_rng();
+    rng.gen::<T>()
 }
